+++ packages/sit-onyx/src/components/OnyxSelectInput/OnyxSelectInput.vue
@@ -124,7 +124,8 @@ const blockTyping = (event: KeyboardEvent) => {
   >
     <OnyxFormElement v-bind="props" :error-messages="errorMessages">
       <template #default="{ id }">
-        <div class="onyx-select-input__wrapper">
+        <!-- eslint-disable-next-line vuejs-accessibility/click-events-have-key-events, vuejs-accessibility/no-static-element-interactions -- Disabled rules are needed here, because of the click event. Otherwise clicking on the padding or gap won't open the select -->
+        <div class="onyx-select-input__wrapper" @click="emit('click')">
           <OnyxLoadingIndicator
             v-if="props.loading"
             class="onyx-select-input__loading"
@@ -152,7 +153,6 @@ const blockTyping = (event: KeyboardEvent) => {
             :value="selectionText"
             :autofocus="props.autofocus"
             autocomplete="off"
-            @click="emit('click')"
             @keydown="blockTyping"
           />
 
@@ -177,7 +177,6 @@ const blockTyping = (event: KeyboardEvent) => {
             :aria-label="t('select.toggleDropDown')"
             tabindex="-1"
             :disabled="props.readonly || props.disabled || props.loading"
-            @click="emit('click')"
           >
             <OnyxIcon :icon="chevronDownUp" />
           </button>