+++ apps/demo-app/CHANGELOG.md
@@ -1,5 +1,26 @@
 # demo-app
 
+## 1.0.0-beta.34
+
+### Patch Changes
+
+- Updated dependencies [934a903]
+  - sit-onyx@1.0.0-beta.34
+
+## 1.0.0-beta.33
+
+### Patch Changes
+
+- Updated dependencies [07084b1]
+  - sit-onyx@1.0.0-beta.33
+
+## 1.0.0-beta.32
+
+### Patch Changes
+
+- Updated dependencies [e8c5341]
+  - sit-onyx@1.0.0-beta.32
+
 ## 1.0.0-beta.31
 
 ### Patch Changes